utility method to extract a pagination cursor from a jsonapi document
