--- patreon/version_compatibility/urlencode.py
@@ -1,6 +1,0 @@
-try:
-    # python2
-    from urllib import urlencode
-except ImportError:
-    # python3
-    from urllib.parse import urlencode
